requirements.txt
wandb~=0.16.3
PyYAML~=6.0
omegaconf~=2.3.0

datasets~=2.18.0
huggingface-hub~=0.21.3
evaluate~=0.4.1
scikit-learn~=1.4.1.post1
numpy~=1.26.4

Pillow~=10.2.0

torch~=2.2.1
accelerate~=0.27.2
torchmetrics~=1.3.1
pytorch-lightning~=2.2.1

tokenizers~=0.19.1
transformers~=4.40.0
